Container refactoring. Finished major part of redesign, created new modules and ContainerBuilder extensions. Replaced usage in DeployPackage and Server applications. Removed server container initialization via web.config.
