+++ Source/Rhetos.Configuration.Autofac/DeployType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rhetos.Configuration.Autofac
+{
+    public enum DeployType
+    {
+        DeployFull,
+        DeployDatabaseOnly
+    }
+}
