+++ Source/Rhetos/Web.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
-    <section name="autofacComponents" type="Autofac.Configuration.SectionHandler, Autofac.Configuration" />
     <section name="nlog" type="NLog.Config.ConfigSectionHandler, NLog" />
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework" />
   </configSections>
@@ -21,11 +20,6 @@
   <entityFramework>
   <queryCache size="10000" cleaningIntervalInSeconds="60" />
   </entityFramework>
-  <autofacComponents>
-    <modules>
-      <module type="Rhetos.DefaultAutofacConfiguration, Rhetos" />
-    </modules>
-  </autofacComponents>
   <nlog throwExceptions="true" xmlns="http://www.nlog-project.org/schemas/NLog.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
     <targets>
       <target name="MainLog" xsi:type="File" fileName="${basedir}\Logs\RhetosServer.log" encoding="utf-8" archiveFileName="${basedir}\Logs\Archives\RhetosServer {#####}.zip" enableArchiveFileCompression="true" archiveAboveSize="2000000" archiveNumbering="DateAndSequence" />
